Adds QueueSizeMatters and it's Template
We were given the task of developing the documentation of a project by our idea. I came up with the idea while thinking about having to wait in a queue in front of the payment bureau at FMI. I liked how cohesively the team worked, even though we had some minor issues with the personality of one of the team members.

The team, which consisted of five members, developed and defended the project within 3 months and got an almost perfect score for it.